requirements.txt
fastapi
uvicorn
transformers
torch
faiss-cpu
numpy
Pillow